pyproject.toml
[tool.poetry]
name = "task4"
version = "0.1.0"
description = "Applying operations for various request types>"
authors = ["Ahmad Tamer"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
pyjwt = "^2.6.0"
datetime = "^4.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
